pyproject.toml
[tool.poetry]
name = "auction"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
psycopg2-binary = "^2.8.6"
Pillow = "^8.1.2"
django-filter = "^2.4.0"
djangorestframework-simplejwt = "^4.6.0"
celery = "^5.0.5"
redis = "^3.5.3"
channels = "^3.0.3"
channels-redis = "^3.2.0"
importlib-metadata = "^4.0.1"
typing-extensions = "^3.10.0"
gunicorn = "^20.1.0"
boto3 = "^1.17.74"
django-storages = "^1.11.1"
daphne = "^3.0.2"
pyparsing = "^3.0.7"
pydot = "^1.4.2"
django-extensions = "^3.1.5"
pytest-cov = "^3.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
